[lightbulbs-for-code-actions] Show lightbulb in modeline to hint at available code actions
Unlike similar features (lsp_auto_highlight_references, automatic
hover information), this is on by default.  The lightbulb should be
unobtrusive enough, and the exta LSP traffic should not matter.

Currently there is no way to prevent kak-lsp from sending code action
requests. We should probably allow that, at least for debugging.

Breaking change: lsp-show-code-action was renamed to
lsp-perform-code-action because lsp-show-code-action now takes the
responsibility to show the lightbulb.
OTOH, lsp-perform-code-action is called when the users requests to run
a code action via lsp-code-actions (bound to "a" in the lsp user mode).

Closes # 475
Closes mawww/kakoune# 4387